requirements.txt
numpy==1.18.1
pandas==1.0.5
pytz==2020.1
requests==2.24.0
websocket-client==0.57.0